requirements.txt
a2a-sdk
uvicorn
httpx
python-dotenv
google-adk>=1.0.0
google-genai>=1.9.0
langchain-google-genai>=1.0.0
pydantic>=2.11.0